requirements.txt
Flask==3.1.2
Flask-RESTX==1.3.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-CORS==6.0.2
PyMySQL==1.1.2
psycopg2-binary==2.9.11
bcrypt==5.0.0
PyJWT==2.11.0
python-dotenv==1.2.1
pandas==3.0.0
numpy==2.4.2
requests==2.32.5
